Save Clerk auth development progress
