+++ package.json
@@ -9,12 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/nextjs": "^5.3.0",
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^5.16.6",
     "@mui/material": "^5.16.6",
-    "firebase": "^10.12.5",
-    "firebaseui": "^6.1.0",
     "next": "14.2.5",
     "openai": "^4.55.1",
     "react": "^18",
